Convert any identifier / file path to Shahar's short id or stack id

anyid2shortid converts any of Shahar's ids (or a filename) to a short id

anyid2stack converts any of Shahar's ids (or a filename) to a stack id

anyid2longid converts any of Shahar's ids (or a filename) to an Igor Pro cell id (i.e. the ones used to name the odour response data in Spikes) object.

anyid2shortid(x)

anyid2stack(x)

anyid2longid(x)

Arguments

x
A stack, cell or short id or a filename that begins with one of these.

Value

A vector of short ids or stack ids, named by the original id when there were more than 1

Details

shortids look like 120726c3

Examples

anyid2shortid("120726BJK1742SF274LC.lsm")

#> [1] "120726c3"

anyid2stack("120726BJK1742SF274LC.lsm")

#> [1] "120726B"

anyid2stack("120726c3")

#> [1] "120726B"

anyid2longid("120726c3")

#> [1] "nm20120726c3"

anyid2longid("120726BJK1742SF274LC.lsm")

#> [1] "nm20120726c3"